ASoC: dapm: Add API to remove widget & it's control dynamically
Currently, asoc provides APIs to add widget & it's controls
dynamically. However, no API exists to remove them.The assumption
is, controls would be removed only in case sound card is not used
and unregistered.

In case of GB codec, we need to add/remove widget & it's controls
dynamically in response to each audio module plugged-in/out.

Signed-off-by: Vaibhav Agarwal <vaibhav.agarwal@linaro.org>